+++ src/mapnik_map.cpp
@@ -46,12 +46,6 @@
 // boost
 #include <boost/foreach.hpp>            // for auto_any_base, etc
 #include <boost/optional/optional.hpp>  // for optional
-#include <boost/smart_ptr/make_shared.hpp>  // for make_shared
-#include <boost/variant/detail/apply_visitor_delayed.hpp>
-#include <boost/variant/detail/apply_visitor_unary.hpp>
-#include <boost/variant/static_visitor.hpp>  // for static_visitor
-#include <boost/variant/variant.hpp>    // for variant
-
 #include "mapnik3x_compatibility.hpp"
 
 Persistent<FunctionTemplate> Map::constructor;
@@ -1447,18 +1441,21 @@ Handle<Value> Map::render(const Arguments& args)
         if (options->Has(String::New("tolerance"))) {
 
             Local<Value> param_val = options->Get(String::New("tolerance"));
-            if (!param_val->IsNumber())
+            if (!param_val->IsNumber()) {
+                delete closure;
                 return ThrowException(Exception::TypeError(
                                           String::New("option 'tolerance' must be an unsigned integer")));
+            }
             closure->tolerance = param_val->IntegerValue();
         }
 
         if (options->Has(String::New("path_multiplier"))) {
-
             Local<Value> param_val = options->Get(String::New("path_multiplier"));
-            if (!param_val->IsNumber())
+            if (!param_val->IsNumber()) {
+                delete closure;
                 return ThrowException(Exception::TypeError(
                                           String::New("option 'path_multiplier' must be an unsigned integer")));
+            }
             closure->path_multiplier = param_val->NumberValue();
         }